Follow similar pattern with validator code.
